nf_automation/requirements.txt
pyautogui==0.9.54
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2
selenium==4.15.2
pywin32==306
